requirements.txt
pre-commit
Pillow~=8.2.0
pymongo~=3.11.4
requests~=2.25.1
Scrapy~=2.5.0
itemadapter~=0.2.0
python-dotenv~=0.17.1
beautifulsoup4~=4.9.3
selenium~=3.141.0
pyvirtualdisplay~=2.1
tabulate
requests
face_recognition
django~=3.2.3
numpy
scipy
matplotlib
scikit-image
scikit-learn
ipython
dlib
facebook-sdk
nltk~=3.6.2
fpdf~=1.7.2
